sdhg

+++ ft_convd.c
@@ -6,7 +6,7 @@
 /*   By: mde-laga <mde-laga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/24 15:43:34 by lramard           #+#    #+#             */
-/*   Updated: 2019/02/11 09:42:44 by mde-laga         ###   ########.fr       */
+/*   Updated: 2019/02/11 12:06:40 by mde-laga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,9 +74,9 @@ static char		*ft_prefix(t_prin *prin, int64_t nb, char *str)
 
 static char		*ft_suffix(t_prin *prin, char *ret)
 {
-	char		*suf;
-	int			i;
-	int64_t	len;
+	char	*suf;
+	int		i;
+	int		len;
 
 	i = 0;
 	len = prin->field - ft_strlen(ret);
